# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "operationId": "be2e78e5f93182d4403f839cc516d312",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pa$$word"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User logged in successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJh..."
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "user@gmail.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+8801712345678",
                                                            "nullable": true
                                                        },
                                                        "profile_photo_url": {
                                                            "type": "string",
                                                            "example": "avatar.jpg",
                                                            "nullable": true
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_admin": {
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "is_property_owner": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_property_inspector": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "example": "2025-06-06 12:00:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials."
                    },
                    "403": {
                        "description": "Your account is suspended."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "422": {
                        "description": "Validation error."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Register",
                "operationId": "d43ecdfd17aca209ecd9b0a8b370e2c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pa$$word"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pa$$word"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJh..."
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "user@gmail.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+8801712345678",
                                                            "nullable": true
                                                        },
                                                        "profile_photo_url": {
                                                            "type": "string",
                                                            "example": "avatar.jpg",
                                                            "nullable": true
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_admin": {
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "is_property_owner": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_property_inspector": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "example": "2025-06-06 12:00:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "operationId": "c360108c781c536eb6f37967647d1fa3",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}